+++ src/PuliApplicationConfig.php
@@ -331,6 +331,7 @@ class PuliApplicationConfig extends DefaultApplicationConfig
                     ->addOption('enabled', null, Option::NO_VALUE, 'Show enabled packages')
                     ->addOption('not-found', null, Option::NO_VALUE, 'Show packages that could not be found')
                     ->addOption('not-loadable', null, Option::NO_VALUE, 'Show packages that could not be loaded')
+                    ->addOption('format', null, Option::REQUIRED_VALUE, 'The format of the output. You can use the variables %name%, %install_path%, %installer% and %state% in the format string', null, 'format')
                     ->setHandlerMethod('handleList')
                 ->end()